+++ ribosomesimulator.h
@@ -26,21 +26,21 @@ class RibosomeSimulator {
   RibosomeSimulator();
   [[nodiscard]] int getState() const;
   void setState(int);
-  void getAlphas(std::vector<double>&, std::vector<int>&);
-  void getDecodingAlphas(std::vector<double>&, std::vector<int>&);
+  void getAlphas(std::vector<float>&, std::vector<int>&);
+  void getDecodingAlphas(std::vector<float>&, std::vector<int>&);
 
-  void setPropensities(std::map<std::string, double>& prop);
-  void setPropensity(std::string&, const double&);
-  double getPropensity(std::string&);
-  void setNonCognate(double noNonCog);
+  void setPropensities(std::map<std::string, float>& prop);
+  void setPropensity(std::string&, const float&);
+  float getPropensity(std::string&);
+  void setNonCognate(float noNonCog);
 
-  std::map<std::string, double> getPropensities();
+  std::map<std::string, float> getPropensities();
   void loadConcentrations(const std::string&);
   void loadConcentrationsFromString(const std::string&);
   void setCodonForSimulation(const std::string&);
-  void run_and_get_times(double&, double&);
-  double run_repeatedly_get_average_time(const int&);
-  std::vector<double> dt_history;
+  void run_and_get_times(float&, float&);
+  float run_repeatedly_get_average_time(const int&);
+  std::vector<float> dt_history;
   std::vector<int> ribosome_state_history;
   std::string saccharomyces_cerevisiae_concentrations = 
        "concentrations/Saccharomyces_cerevisiae.csv";
@@ -61,13 +61,13 @@ class RibosomeSimulator {
   void buildReactionsMap();
   std::string simulation_codon_3_letters;
   csv_utils::ConcentrationsReader concentrations_reader;
-  std::vector<std::vector<std::tuple<std::reference_wrapper<double>, int>>>
+  std::vector<std::vector<std::tuple<std::reference_wrapper<float>, int>>>
   createReactionsGraph(const csv_utils::concentration_entry&);
   std::map<
       std::string,
-      std::vector<std::vector<std::tuple<std::reference_wrapper<double>, int>>>>
+      std::vector<std::vector<std::tuple<std::reference_wrapper<float>, int>>>>
       reactions_map;
-  std::vector<std::vector<std::tuple<std::reference_wrapper<double>, int>>>
+  std::vector<std::vector<std::tuple<std::reference_wrapper<float>, int>>>
       reactions_graph;  // vector where the index is the ribosome's current
                         // state and the content is a vector of tuples
                         // containing the propensity and next state of each
@@ -76,68 +76,68 @@ class RibosomeSimulator {
 
   std::vector<std::string> stop_codons = {"UAG", "UAA", "UGA"};
   // constants for WCcognate interaction in 1/sec
-  std::map<std::string, double> WC1f;
-  double WC1r = 85;
-  double WC2f = 190;
-  double WC2r = 0.23;
-  double WC3f = 260;
-  double WC4f = 1000;
-  double WC5f = 1000;
-  double WCdiss = 60;
-  double WC6f = 1000;
-  double dec7f = 200;
+  std::map<std::string, float> WC1f;
+  float WC1r = 85;
+  float WC2f = 190;
+  float WC2r = 0.23;
+  float WC3f = 260;
+  float WC4f = 1000;
+  float WC5f = 1000;
+  float WCdiss = 60;
+  float WC6f = 1000;
+  float dec7f = 200;
 
   // constants for wobblecognate interaction in 1/sec
-  std::map<std::string, double> wobble1f;
-  double wobble1r = 85;
-  double wobble2f = 190;
-  double wobble2r = 1;
-  double wobble3f = 25;
-  double wobble4f = 1000;
-  double wobble5f = 1000;
-  double wobblediss = 1.1;
-  double wobble6f = 6.4;
+  std::map<std::string, float> wobble1f;
+  float wobble1r = 85;
+  float wobble2f = 190;
+  float wobble2r = 1;
+  float wobble3f = 25;
+  float wobble4f = 1000;
+  float wobble5f = 1000;
+  float wobblediss = 1.1;
+  float wobble6f = 6.4;
 
   // constants for nearcognate interaction in 1/sec
-  std::map<std::string, double> near1f;
-  double near1r = 85;
-  double near2f = 190;
-  double near2r = 80;
-  double near3f = 0.4;
-  double near4f = 1000;
-  double near5f = 1000;
-  double neardiss = 1000;
-  double near6f = 60;
+  std::map<std::string, float> near1f;
+  float near1r = 85;
+  float near2f = 190;
+  float near2r = 80;
+  float near3f = 0.4;
+  float near4f = 1000;
+  float near5f = 1000;
+  float neardiss = 1000;
+  float near6f = 60;
 
-  double totalconc = 1.9e-4;
+  float totalconc = 1.9e-4;
 
   // constants for noncognate interaction in 1/sec.
   // Non-cognates are assumed to not undergo any significant
   // interaction but to simply dissociate quickly.
-  std::map<std::string, double> non1f;
-  double non1r = 1e5;
+  std::map<std::string, float> non1f;
+  float non1r = 1e5;
 
   // based on yeast value of 226000 molecules per cell as determined
   // in von der Haar 2008 (PMID 18925958)
-  double eEF2conc = 1.36e-5;
+  float eEF2conc = 1.36e-5;
   // constants for translocation in 1/sec
   // 150 uM-1 s-1 = is from Fluitt et al 2007 (PMID 17897886)
-  double trans1f = eEF2conc * 1.5e8;
-  double trans1r = 140;
-  double trans2 = 250;
-  double trans3 = 350;
-  double trans4 = 1000;
-  double trans5 = 1000;
-  double trans6 = 1000;
-  double trans7 = 1000;
-  double trans8 = 1000;
-  double trans9 = 1000;
+  float trans1f = eEF2conc * 1.5e8f;
+  float trans1r = 140;
+  float trans2 = 250;
+  float trans3 = 350;
+  float trans4 = 1000;
+  float trans5 = 1000;
+  float trans6 = 1000;
+  float trans7 = 1000;
+  float trans8 = 1000;
+  float trans9 = 1000;
 
   // TEMPORARY SOLUTION: in order to provide a mechanism easily change only one
   // propensity I've created a map of references to the propensities variables,
   // but this must be updated manually if in the future we add or remove
   // reactions.
-  std::map<std::string, double*> propensities_map;
+  std::map<std::string, float*> propensities_map;
 };
 
 }  // namespace Simulations
